+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES for `command-line-publish`
 
-## ?
+## 0.2.0
+
+- Refactor: Use `pkg.name` for header (better reusability across projects)
+- Refactor: Use external `command-line-basics` for help, version, notifier
+
+## 0.1.0
 
 - Initial version